Add back button to recipe pages
A back button is an essential part of every website. With this commit,
all the recipe pages receive a back button, which redirects the user
back to index.html.

A back button makes it easier for the user to navigate through separate
pages without having to use the browser back button, which can sometimes
lead to unexpected errors.