Style main page
Main page gets styles with CSS. Specifically, the background is changed
to dark, the title is centered and bigger, all texts are also bigger,
and there is an image along with the text for all recipes.

An image allows for the user to visualize what they may be learning
about, even before clicking on the page. Without an image, the web page
will look bland, as it includes a mix of only 3 colors.